LAB-5/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
openai==1.40.0
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.8.0
redis==5.0.4
python-multipart==0.0.9
aiohttp==3.9.5
pydantic-settings==2.3.4
